wibdns/requirements.txt
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
TgCrypto==1.2.5
meval==2.5
aiosqlite==0.20.0
aiofiles==23.2.1
aiocsv==1.3.2
Hydrogram==0.2.0
uvloop==0.19.0
paramiko==3.4.1